requirements.txt
streamlit==1.45.1
opencv-python-headless==4.11.0.86
torch==2.7.1
torchvision==0.22.1
Pillow==10.0.0
numpy==1.24.3